requirements.txt
fastapi
uvicorn
langchain
langchain-groq
langgraph
tavily-python
newspaper3k
python-dotenv
langchain_community
lxml_html_clean